move stub to a file
