+++ wire-ios-request-strategy/Tests/Sources/Payloads/Processing/ConversationEventPayloadProcessorTests.swift
@@ -1013,46 +1013,3 @@ final class ConversationEventPayloadProcessorTests: MessagingTestBase {
     }
 
 }
-
-extension Payload.Conversation {
-
-    static func stub(
-        qualifiedID: QualifiedID? = nil,
-        id: UUID?  = nil,
-        type: BackendConversationType? = nil,
-        creator: UUID? = nil,
-        access: [String]? = nil,
-        accessRole: String? = nil,
-        accessRoles: [String]? = nil,
-        name: String? = nil,
-        members: Payload.ConversationMembers? = nil,
-        lastEvent: String? = nil,
-        lastEventTime: String? = nil,
-        teamID: UUID? = nil,
-        messageTimer: TimeInterval? = nil,
-        readReceiptMode: Int? = nil,
-        messageProtocol: String? = "proteus",
-        mlsGroupID: String? = "id".data(using: .utf8)?.base64EncodedString()
-    ) -> Self {
-
-        self.init(
-            qualifiedID: qualifiedID,
-            id: id,
-            type: type?.rawValue,
-            creator: creator,
-            access: access,
-            legacyAccessRole: accessRole,
-            accessRoles: accessRoles,
-            name: name,
-            members: members,
-            lastEvent: lastEvent,
-            lastEventTime: lastEventTime,
-            teamID: teamID,
-            messageTimer: messageTimer,
-            readReceiptMode: readReceiptMode,
-            messageProtocol: messageProtocol,
-            mlsGroupID: mlsGroupID
-        )
-
-    }
-}
